requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchmetrics>=1.0.0

# ONNX 관련
onnx>=1.14.0
onnxruntime>=1.15.0
onnxruntime-gpu>=1.15.0

timm>=0.9.0
Pillow>=9.0.0

pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0

fastapi>=0.100.0
uvicorn>=0.20.0
python-multipart>=0.0.6
psutil>=5.9.0
requests>=2.31.0

mlflow>=2.5.0
tensorboard>=2.12.0

PyYAML>=6.0

tqdm>=4.64.0
matplotlib>=3.5.0
seaborn>=0.11.0

pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

pydicom>=2.3.0
opencv-python>=4.6.0

loguru>=0.6.0 

pytest==7.4.0
httpx>=0.24.0